update examples to use new DLML method names and skip auto pop files
The following examples have been updated and tested:
1-9, 14-16
