+++ E16ElectricPowerEarthquake/input.json
@@ -22,12 +22,12 @@
             "PowerNetwork": {
                 "Application": "Pelicun3",
                 "ApplicationData": {
-                    "DL_Method": "HAZUS MH EQ IM",
+                    "DL_Method": "Hazus Earthquake - Electric Power",
                     "Realizations": 5,
-                    "auto_script": "PelicunDefault/Hazus_Earthquake_IM.py",
                     "coupled_EDP": true,
                     "detailed_results": false,
                     "ground_failure": false,
+                    "lifeline_facility": false,
                     "log_file": true,
                     "regional": "true"
                 }
