+++ E1BasicHAZUS/input.json
@@ -4,7 +4,7 @@
             "Buildings": {
                 "Application": "CSV_to_AIM",
                 "ApplicationData": {
-                    "assetSourceFile": "SanFrancisco_buildings.csv",
+                    "assetSourceFile": "SanFrancisco_buildings_ext.csv",
                     "filter": "8000-8024",
                     "inputIsGIS": false,
                     "pathToSource": "{Current_Dir}/input_data"
@@ -15,12 +15,12 @@
             "Buildings": {
                 "Application": "Pelicun3",
                 "ApplicationData": {
-                    "DL_Method": "HAZUS MH EQ IM",
+                    "DL_Method": "Hazus Earthquake - Buildings",
                     "Realizations": 100,
-                    "auto_script": "PelicunDefault/Hazus_Earthquake_IM.py",
                     "coupled_EDP": true,
                     "detailed_results": false,
                     "ground_failure": false,
+                    "lifeline_facility": true,
                     "log_file": true,
                     "regional": "true"
                 }
